Update Flysystem isDir/isFile to return false for meta errors.

+++ src/Sync/Task/CheckFolderPlaylistsTask.php
@@ -7,7 +7,6 @@ use App\Entity;
 use App\Flysystem\StationFilesystems;
 use Azura\Files\ExtendedFilesystemInterface;
 use Doctrine\ORM\Query;
-use League\Flysystem\UnableToRetrieveMetadata;
 use Psr\Log\LoggerInterface;
 
 class CheckFolderPlaylistsTask extends AbstractTask
@@ -104,9 +103,7 @@ class CheckFolderPlaylistsTask extends AbstractTask
             $path = $folder->getPath();
 
             // Verify the folder still exists.
-            try {
-                $fsMedia->isDir($path);
-            } catch (UnableToRetrieveMetadata $exception) {
+            if (!$fsMedia->isDir($path)) {
                 $this->em->remove($folder);
                 continue;
             }
